requirements.txt
requests==2.31.0
pymongo==4.6.0
pandas==2.2.2
matplotlib==3.9.0
python-dotenv==1.0.1
jupyter==1.0.0
scipy==1.11.0
ipykernel==6.29.0
